extracting Background Video to atoms and creating base files of the main section in organism
