+++ src/components/atoms/BgVideo/BgVideo.styles.tsx
@@ -1,9 +1,4 @@
-import styled from 'styled-components/macro';
-
-export const Header = styled.header`
-    position: relative;
-    height: 100vh;
-`;
+import styled from "styled-components";
 
 export const BgVideo = styled.div`
     position: absolute;
@@ -20,4 +15,3 @@ export const BgVideoContent = styled.video`
     height: 100%;
     object-fit: cover;
 `;
-
