+++ src/components/atoms/BgVideo/BgVideo.tsx
@@ -1,15 +1,12 @@
 import React from 'react';
 import videoHeaderMp4 from 'assets/video/CuttingWoodmp4.mp4';
 import videoHeaderWebm from 'assets/video/CuttingWoodWebm.webm';
-import Logo from 'components/atoms/Logo/Logo';
-import { Heading } from 'components/atoms/HeadingBox/Heading';
-import { BgVideoContent, BgVideo, Header } from './Header.styles';
+import { BgVideo, BgVideoContent } from "./BgVideo.styles";
 
-export const HeaderWrapper = (): JSX.Element => {
+export const BgVideoWrapper = (): JSX.Element => {
     const noVideoInfo = 'Your browser does not support video';
 
     return (
-        <Header>
             <BgVideo>
                 <BgVideoContent autoPlay muted loop>
                     <source src={videoHeaderMp4} />
@@ -17,8 +14,5 @@ export const HeaderWrapper = (): JSX.Element => {
                     {noVideoInfo}
                 </BgVideoContent>
             </BgVideo>
-            <Heading />
-            <Logo />
-        </Header>
     );
 };
