+++ src/components/organism/Header/Header.styles.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components/macro';
+
+export const Header = styled.header`
+    position: relative;
+    height: 100vh;
+`;
+
