+++ src/components/organism/SectionAbout/MainSection.styles.tsx
@@ -1,0 +1,1 @@
+import styled from 'styled-components/macro';
