+++ src/components/organism/SectionAbout/MainSection.tsx
@@ -1,0 +1,1 @@
+import React from "react";
