+++ src/views/App.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
-import { HeaderWrapper } from 'components/organism/Header';
+import { HeaderWrapper } from 'components/organism/Header/Header';
 import Navigation from 'components/molecules/Navigation/Navigation';
+
 
 const App = (): JSX.Element => {
     return (
